backend/requirements.txt
python-dotenv==1.0.1
black==24.8.0
isort==5.13.2
langchain==0.2.14
langchain-openai==0.1.22
langchain-community==0.2.12
langchain-core==0.2.33
langgraph==0.2.4
langchain-groq==0.1.9
fastapi==0.112.1
uvicorn==0.30.6
pillow==10.4.0
langchain-anthropic==0.1.23
python-multipart==0.0.9
